+++ src/state/errorTextState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const errorTextState = atom<string>({
+  key: 'errorTextState',
+  default: '',
+})
