+++ src/state/isErrorState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const isErrorState = atom<boolean>({
+  key: 'isErrorState',
+  default: false,
+})
